Lab4/worker/requirements.txt
Flask>=2.0.0
gunicorn>=20.0.0
kafka-python>=2.0.0
requests>=2.30.0
